+++ RELEASES.md
@@ -2,6 +2,8 @@
 
 #### 27.1.1
   - New: Auto-detect the correct FTP listing parser when SYST command fails (IIS, Azure, OpenVMS)
+  - New: Assume FTP commands supported by OpenVMS HGFTP server
+  - FiX: Support edge case for OpenVMS absolute paths (directive can be alpha-numeric)
 
 #### 27.1.0
   - New: Improved transfer rate throttling when using an upload/download speed limit (thanks [wakabayashik](/wakabayashik))